+++ mod-source-record-storage-server/src/main/java/org/folio/services/util/parser/lexeme/operand/BinaryOperand.java
@@ -3,6 +3,8 @@ package org.folio.services.util.parser.lexeme.operand;
 import org.folio.services.util.parser.lexeme.Lexeme;
 import org.folio.services.util.parser.lexeme.Lexicon;
 
+import java.util.List;
+
 /**
  * The interface for the binary operands. It has a key(field.suffix), operator, and right operand (value).
  * @see Lexeme
@@ -12,5 +14,5 @@ public interface BinaryOperand {
 
   String getField();
 
-  String getBindingParam();
+  List<String> getBindingParams();
 }